Fix some potential crashing, refactoring and adding C-d/C-u
